refactor: Extract view module by Blueprint - Move the view function to a separate Blueprint module - Clean up main.py
